Fixed compile warnings

+++ include/terrau/binfile.h
@@ -9,7 +9,7 @@ extern "C" {
 #endif
 
 terra_status_t terrau_read_binary_file(
-    terra_app_t *app, const char *filename, int64_t *out_size, char **out
+    terra_app_t *app, const char *filename, size_t *out_size, char **out
 );
 terra_status_t terrau_visualize_binary_file(
     terra_app_t *app, const char *contents, const int64_t size
